Feat: Finishing touches
I have updated the ReadMe to make it easier for people to understand the setup process for both the front and backend.
I have also updated the API home view request in order to help show the commands you can do with the API and to better understand the process.